[4269] Moving profile activation to pom/mavensettings where
we can look at other profiles state in order to determine what to do
with `activeByDefault` profiles.

I'm not happy that I've copied the code in multiple locations.  The
method of isolating code to a static method doesn't work so well and my
attempts at doing some interface/inheritance modeling is not working
so well with lombok, which I've never really worked with.

Lastly, I've added some tests related to profile deactivation.  If
there's code that handles that, I haven't found it.  The code I've
added here does not.

 #4269