[4269] An interface-based way to isolate profile code
Honestly, I'd probably just copy the code for now.  If it weren't for
having to have listProfiles due to lombok-generated
`MavenSettings#getProfile`, which has a different signature than we
need, I'd probably go with this.  But having `getProfiles` and
`listProfiles` in the same class is really messy.

Even so, I'll push it and let the rewrite folks decide.

 #4269